Numeric.RootFinding

Description

Haskell functions for finding the roots of real functions of real arguments. These algorithms are iterative so we provide both function returning root (or failure to find root) and list of iterations.

Data types

data Root a

The result of searching for a root of a mathematical function.

Constructors

NotBracketed
The function does not have opposite signs when evaluated at the lower and upper bounds of the search.

SearchFailed
The search failed to converge to within the given error tolerance after the given number of iterations.

Root !a
A root was successfully found.

fromRoot

Arguments

:: a
Default value.

-> Root a
Result of search for a root.

-> a

Returns either the result of a search for a root, or the default value if the search failed.

data Tolerance

Error tolerance for finding root. It describes when root finding algorithm should stop trying to improve approximation.

Constructors

RelTol !Double
Relative error tolerance. Given RelTol ε two values are considered approximately equal if \[ \frac{|a - b|}{|\operatorname{max}(a,b)} < \varepsilon \]

AbsTol !Double
Absolute error tolerance. Given AbsTol δ two values are considered approximately equal if \[ |a - b| < \delta \]. Note that AbsTol 0 could be used to require to find approximation within machine precision.

withinTolerance :: Tolerance -> Double -> Double -> Bool

Check that two values are approximately equal. In addition to specification values are considered equal if they're within 1ulp of precision. No further improvement could be done anyway.

class IterationStep a where

Type class for checking whether iteration converged already.

Methods

matchRoot :: Tolerance -> a -> Maybe (Root Double)

Return Just root is current iteration converged within required error tolerance. Returns Nothing otherwise.

findRoot

Arguments

:: IterationStep a

=> Int
Maximum

-> Tolerance
Error tolerance

-> [a]

-> Root Double

Find root in lazy list of iterations.

Ridders algorithm

data RiddersParam

Parameters for ridders root finding

Constructors

RiddersParam

Fields

riddersMaxIter :: !Int

Maximum number of iterations. Default = 100

riddersTol :: !Tolerance

Error tolerance for root approximation. Default is relative error 4·ε, where ε is machine precision.

ridders

Arguments

:: RiddersParam
Parameters for algorithms. def provides reasonable defaults

-> (Double, Double)
Bracket for root

-> (Double -> Double)
Function to find roots

-> Root Double

Use the method of Ridders[Ridders1979] to compute a root of a function. It doesn't require derivative and provide quadratic convergence (number of significant digits grows quadratically with number of iterations).

The function must have opposite signs when evaluated at the lower and upper bounds of the search (i.e. the root must be bracketed). If there's more that one root in the bracket iteration will converge to some root in the bracket.

riddersIterations :: (Double, Double) -> (Double -> Double) -> [RiddersStep]

List of iterations for Ridders methods. See ridders for documentation of parameters

data RiddersStep

Single Ridders step. It's a bracket of root

Constructors

RiddersStep !Double !Double
Ridders step. Parameters are bracket for the root

RiddersBisect !Double !Double
Bisection step. It's fallback which is taken when Ridders update takes us out of bracket

RiddersRoot !Double
Root found

RiddersNoBracket
Root is not bracketed

Newton-Raphson algorithm

data NewtonParam

Parameters for ridders root finding

Constructors

NewtonParam

Fields

newtonMaxIter :: !Int

Maximum number of iterations. Default = 50

newtonTol :: !Tolerance

Error tolerance for root approximation. Default is relative error 4·ε, where ε is machine precision

newtonRaphson

Arguments

:: NewtonParam
Parameters for algorithm. def provide reasonable defaults.

-> (Double, Double, Double)
Triple of (low bound, initial guess, upper bound). If initial guess if out of bracket middle of bracket is taken as approximation

-> (Double -> (Double, Double))
Function to find root of. It returns pair of function value and its first derivative

-> Root Double

Solve equation using Newton-Raphson iterations.

This method require both initial guess and bounds for root. If Newton step takes us out of bounds on root function reverts to bisection.

newtonRaphsonIterations :: (Double, Double, Double) -> (Double -> (Double, Double)) -> [NewtonStep]

List of iteration for Newton-Raphson algorithm. See documentation for newtonRaphson for meaning of parameters.

data NewtonStep

Steps for Newton iterations

Constructors

NewtonStep !Double !Double
Normal Newton-Raphson update. Parameters are: old guess, new guess

NewtonBisection !Double !Double
Bisection fallback when Newton-Raphson iteration doesn't work. Parameters are bracket on root

NewtonRoot !Double
Root is found

NewtonNoBracket
Root is not bracketed

References

Ridders, C.F.J. (1979) A new algorithm for computing a single root of a real continuous function. IEEE Transactions on Circuits and Systems 26:979–980.

Press W.H.; Teukolsky S.A.; Vetterling W.T.; Flannery B.P. (2007). "Section 9.2.1. Ridders' Method". /Numerical Recipes: The Art of Scientific Computing (3rd ed.)./ New York: Cambridge University Press. ISBN 978-0-521-88068-8.