Numeric.MathFunctions.Comparison

Description

Functions for approximate comparison of floating point numbers.

Approximate floating point comparison, based on Bruce Dawson's "Comparing floating point numbers": http://www.cygnus-software.com/papers/comparingfloats/comparingfloats.htm

Relative erros

relativeError :: Double -> Double -> Double

Calculate relative error of two numbers:

|a - b| / max |a| |b|

It lies in [0,1) interval for numbers with same sign and (1,2] for numbers with different sign. If both arguments are zero or negative zero function returns 0. If at least one argument is transfinite it returns NaN

eqRelErr

Arguments

:: Double
eps relative error should be in [0,1) range

-> Double
a

-> Double
b

-> Bool

Check that relative error between two numbers a and b. If relativeError returns NaN it returns False.

Ulps-based comparison

addUlps :: Int -> Double -> Double

Add N ULPs (units of least precision) to Double number.

ulpDistance :: Double -> Double -> Word64

Measure distance between two Doubles in ULPs (units of least precision).

within

Arguments

:: Int
Number of ULPs of accuracy desired.

-> Double

-> Double

-> Bool

Compare two Double values for approximate equality, using Dawson's method.

The required accuracy is specified in ULPs (units of least precision). If the two numbers differ by the given number of ULPs or less, this function returns True.