Numeric.RootFinding

Description

Haskell functions for finding the roots of real functions of real arguments.

Documentation

data Root a

The result of searching for a root of a mathematical function.

Constructors

NotBracketed
The function does not have opposite signs when evaluated at the lower and upper bounds of the search.

SearchFailed
The search failed to converge to within the given error tolerance after the given number of iterations.

Root a
A root was successfully found.

fromRoot

Arguments

:: a
Default value.

-> Root a
Result of search for a root.

-> a

Returns either the result of a search for a root, or the default value if the search failed.

ridders

Arguments

:: Double
Absolute error tolerance.

-> (Double, Double)
Lower and upper bounds for the search.

-> (Double -> Double)
Function to find the roots of.

-> Root Double

Use the method of Ridders to compute a root of a function.

The function must have opposite signs when evaluated at the lower and upper bounds of the search (i.e. the root must be bracketed).

newtonRaphson

Arguments

:: Double
Required precision

-> (Double, Double, Double)
(lower bound, initial guess, upper bound). Iterations will no go outside of the interval

-> (Double -> (Double, Double))
Function to finds roots. It returns pair of function value and its derivative

-> Root Double

Solve equation using Newton-Raphson iterations.

This method require both initial guess and bounds for root. If Newton step takes us out of bounds on root function reverts to bisection.

References

Ridders, C.F.J. (1979) A new algorithm for computing a single root of a real continuous function. IEEE Transactions on Circuits and Systems 26:979–980.